refactor: Move utils into duck folder
