+++ src/ducks/onboarding/EnsureHasAccounts.jsx
@@ -6,7 +6,7 @@ import Onboarding from './Onboarding'
 import styles from './Onboarding.styl'
 import stylesLayout from 'styles/main.styl'
 import { Layout, Main, Content } from 'cozy-ui/react'
-import { queryConnect, isCollectionLoading } from 'utils/client'
+import { queryConnect, isCollectionLoading } from 'ducks/client/utils'
 import { getClient } from 'ducks/client'
 import { accountsConn } from 'doctypes'
