+++ src/ducks/settings/AccountsSettings.jsx
@@ -7,7 +7,7 @@ import { groupBy, flowRight as compose, sortBy } from 'lodash'
 import { getAppUrlById } from 'selectors'
 import Table from 'components/Table'
 import Loading from 'components/Loading'
-import { queryConnect } from 'utils/client'
+import { queryConnect } from 'ducks/client/utils'
 import plus from 'assets/icons/16/plus.svg'
 import styles from './AccountsSettings.styl'
 import btnStyles from 'styles/buttons.styl'
