+++ src/ducks/settings/Configuration.jsx
@@ -1,7 +1,7 @@
 import React, { Component } from 'react'
 import TogglePane from './TogglePane'
 import { translate } from 'cozy-ui/react'
-import { queryConnect, isCollectionLoading } from 'utils/client'
+import { queryConnect, isCollectionLoading } from 'ducks/client/utils'
 import { settingsConn } from 'doctypes'
 import { flowRight as compose } from 'lodash'
 import Loading from 'components/Loading'
